backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
boto3==1.35.67
botocore==1.35.67
certifi==2024.8.30
click==8.1.7
dnspython==2.7.0
ecdsa==0.19.0
fastapi==0.115.5
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
jmespath==1.0.1
packaging==24.2
pluggy==1.5.0
pyasn1==0.6.1
pydantic==2.10.1
pydantic-settings==2.6.1
pydantic_core==2.27.1
pymongo==4.10.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.17
PyYAML==6.0.2
rsa==4.9
s3transfer==0.10.4
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==0.24.0
websockets==14.1
